Using xsl:variable in xmlPollTransform.xml
The variable has to be passed and enclosed in curly brackets, within an html
form element e.g <input type="radio" name="{$id}", the variable name
should also be prefixed with the $ sign. I am yet to test this by using
PHP on the server side and storing the values on the database.

!!!!!!!!